+++ ext/java/nokogiri/XmlReader.java
@@ -140,14 +140,6 @@ public class XmlReader extends RubyObject
 
   @JRubyMethod
   public IRubyObject
-  attribute_nodes(ThreadContext context)
-  {
-    context.runtime.getWarnings().warn("Reader#attribute_nodes is deprecated and will be removed in a future version of Nokogiri. Please use Reader#attribute_hash instead.");
-    return currentNode().getAttributesNodes();
-  }
-
-  @JRubyMethod
-  public IRubyObject
   attribute_hash(ThreadContext context)
   {
     return currentNode().getAttributes(context);