+++ ext/java/nokogiri/XmlSchema.java
@@ -154,8 +154,7 @@ public class XmlSchema extends RubyObject
       throw context.runtime.newTypeError(msg);
     }
     if (!(rbDocument instanceof XmlDocument)) {
-      // TODO: deprecate allowing Node
-      context.runtime.getWarnings().warn("Passing a Node as the first parameter to Schema.from_document is deprecated. Please pass a Document instead. This will become an error in a future release of Nokogiri.");
+      context.runtime.getWarnings().warn("Passing a Node as the first parameter to Schema.from_document is deprecated. Please pass a Document instead. This will become an error in Nokogiri v1.17.0."); // TODO: deprecated in v1.15.3, remove in v1.17.0
     }
 
     XmlDocument doc = ((XmlDocument)((XmlNode) rbDocument).document(context));